# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Intelligent AI Assistant API
  description: |
    Next-generation **Agentic AI Life Assistant** API designed to act as the user's conscious digital twin, 
    continuously learning, guiding, and optimizing real-world actions in real time.
    
    ## Features
    - **Voice Processing**: Speech-to-text and text-to-speech with Whisper and Coqui/ElevenLabs
    - **Calendar Integration**: Google Calendar synchronization with intelligent scheduling
    - **WhatsApp Messaging**: Business Cloud API integration for confirmations and updates
    - **Federated Learning**: Privacy-preserving machine learning with FedAvg and Differential Privacy
    - **Digital Avatar**: 2D/3D animated user representation during interactions
    - **Privacy Compliance**: GDPR, DIFC, and UAE Data Protection Law compliance
    
    ## Authentication
    All endpoints (except health checks and webhook verification) require JWT authentication.
    Include the Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - **Authentication endpoints**: 5 requests per minute
    - **Voice processing**: 20 requests per minute
    - **General API**: 100 requests per minute
    
    ## Error Handling
    The API uses standard HTTP status codes and returns structured error responses:
    ```json
    {
      "error": "Error description",
...
      "request_id": "uuid",
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /calendar/connect:
    post:
      tags: [Calendar Integration]
      summary: Initiate Google Calendar connection
      description: |
        Start OAuth2 flow for Google Calendar integration.
        Returns authorization URL for user to grant permissions.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                redirect_uri:
                  type: string
                  format: uri
                  description: Custom redirect URI (optional)
                  example: "https://myapp.com/calendar/callback"
      responses:
        '200':
          description: Authorization URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorization_url:
                    type: string
                    format: uri
                    example: "https://accounts.google.com/o/oauth2/auth?client_id=..."
                  state:
                    type: string
                    example: "csrf_token_123"
                  redirect_uri:
                    type: string
                    format: uri
                    example: "https://api.ai-assistant.com/api/v1/calendar/oauth/callback"
  /calendar/events:
    get:
      tags: [Calendar Integration]
      summary: Get calendar events
      description: Retrieve user's calendar events with optional date filtering
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
          description: Start date filter (ISO 8601)
          example: "2024-01-01T00:00:00Z"
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
          description: End date filter (ISO 8601)
          example: "2024-01-31T23:59:59Z"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          description: Maximum number of events to return
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CalendarEvent'

    post:
      tags: [Calendar Integration]
      summary: Create calendar event
      description: Create a new calendar event with AI assistance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarEventCreate'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEvent'

  # WhatsApp Integration Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.
        
        **Token Format:**
        ```
        Authorization: Bearer <jwt_token>
        ```
        
        **Token Lifecycle:**
        - Access tokens expire after 1 hour
        - Refresh tokens expire after 30 days
        - Automatic token rotation supported
  schemas:
...
    CalendarEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        google_event_id:
          type: string
          nullable: true
          example: "google_event_123"
        title:
          type: string
          example: "Team Meeting"
        description:
          type: string
          nullable: true
          example: "Weekly team standup meeting"
        start_time:
          type: string
          format: date-time
          example: "2024-01-15T10:00:00Z"
        end_time:
          type: string
          format: date-time
          example: "2024-01-15T11:00:00Z"
        location:
          type: string
          nullable: true
          example: "Conference Room A"
        attendees:
          type: array
          items:
            type: string
            format: email
          example: ["john@example.com", "jane@example.com"]
        ai_generated:
          type: boolean
          description: Whether event was created by AI
          example: true
    CalendarEventCreate:
      type: object
      required: [title, start_time, end_time]
      properties:
        title:
          type: string
          maxLength: 500
          example: "Team Meeting"
        description:
          type: string
          nullable: true
          example: "Weekly team standup meeting"
        start_time:
          type: string
          format: date-time
          example: "2024-01-15T10:00:00Z"
        end_time:
          type: string
          format: date-time
          example: "2024-01-15T11:00:00Z"
        location:
          type: string
          nullable: true
          example: "Conference Room A"
        attendees:
          type: array
          items:
            type: string
            format: email
          example: ["john@example.com", "jane@example.com"]
...